+++ packages/plugin-packages/mashroom-session-provider-redis/src/mashroom-bootstrap.ts
@@ -10,8 +10,14 @@ import type {IORedisConfig} from '../type-definitions';
 const bootstrap: MashroomSessionStoreProviderPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder, expressSession) => {
     const {loggerFactory, services: {core: {pluginService, healthProbeService}}} =  pluginContextHolder.getPluginContext();
     const logger = loggerFactory('mashroom.session.provider.redis');
+    const redisConfig = pluginConfig as IORedisConfig;
 
-    await setConfig(pluginConfig as IORedisConfig);
+    // We have to move the keyPrefix to RedisStore,
+    // otherwise counting and destroying of sessions won't work
+    const keyPrefix = redisConfig.redisOptions.keyPrefix;
+    delete redisConfig.redisOptions.keyPrefix;
+
+    await setConfig(redisConfig);
     const client = await createClient(logger);
 
     healthProbeService.registerProbe(pluginName, healthProbe);
@@ -27,6 +33,7 @@ const bootstrap: MashroomSessionStoreProviderPluginBootstrapFunction = async (pl
     const RedisStore = createRedisStore(expressSession);
     return new RedisStore({
         client,
+        prefix: keyPrefix,
     });
 };